Re #244: Don't use Entity Row and Rowset as the default entity.
KModelEntityDefault now instantiates the generic KModelEntityComposite for collections and KModelEntityDefault for a resource.

KModelDatabase directly creates rows and rowsets so there seems to be no need to make row/rowset as defaults especially when using a non-database model.

This issue came up when creating a generic web service/resources model. If the default entities are row/rowset, we have to define every entity that are not either row or rowset.