Re #248: Created the Options API
To Load or Create a new Options Object:

$manager->getObject('lib:options.provider')->load('identifier');

KOptions implements KOptionsInterface with these methods: getIdentifier(), getProvider(), and store().

The reason why those are expected is to allow the KOptions to "store" itself using the Provider. When the Provider creates an Options object, it injects the identifier and itself into the Options object. KOptions can't be instantiated without a Provider and Identifier.